Muchas cosas. Proyecto cerca del fin a falta de pruebas y de revisar algunos casos mas de INSERT. Importante arreglar el set order, añadir que los campos de formulario se añadan al payment details

+++ src/main/java/com/iyg16260/farmasterrae/dto/products/ProductDTO.java
@@ -2,12 +2,14 @@ package com.iyg16260.farmasterrae.dto.products;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductDTO {
     String name;
     String description;
     String imagePath;
-    Double price;
+    BigDecimal price;
     Integer stock;
     String reference;
 }
