+++ src/main/java/com/iyg16260/farmasterrae/model/Product.java
@@ -32,7 +32,7 @@ public class Product {
         }
     }
 
-    Double price;
+    Integer price;
 
     Integer stock;
